pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "flask>=3.1.3",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pandas-ta>=0.4.71b0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "requests>=2.33.0",
    "yfinance>=1.2.0",
]
